requirements.txt
streamlit==0.57.3
seaborn==0.10.0
pandas==1.0.5
numpy==1.18.1
matplotlib==3.2.2
Pillow==7.1.2
